fixed edit link
