+++ content/.vuepress/theme/layouts/Post.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import { normalize, outboundRE, endingSlashRE } from "../util";
+import { createEditLink } from "../util";
 
 export default {
     props: ["sidebarItems"],
@@ -57,14 +57,8 @@ export default {
                 docsRepo = repo
             } = this.$site.themeConfig;
 
-            let path = normalize(this.$page.path);
-            if (endingSlashRE.test(path)) {
-                path += "README.md";
-            } else {
-                path += ".md";
-            }
             if (docsRepo && editLinks) {
-                return this.createEditLink(repo, docsRepo, docsDir, docsBranch, path);
+                return createEditLink(repo, docsRepo, docsDir, docsBranch, this.$page.path);
             }
         },
 
@@ -76,33 +70,6 @@ export default {
             );
         }
     },
-
-    methods: {
-        createEditLink(repo, docsRepo, docsDir, docsBranch, path) {
-            const bitbucket = /bitbucket.org/;
-            if (bitbucket.test(repo)) {
-                const base = outboundRE.test(docsRepo) ? docsRepo : repo;
-                return (
-                    base.replace(endingSlashRE, "") +
-                    `/${docsBranch}` +
-                    (docsDir ? "/" + docsDir.replace(endingSlashRE, "") : "") +
-                    path +
-                    `?mode=edit&spa=0&at=${docsBranch}&fileviewer=file-view-default`
-                );
-            }
-
-            const base = outboundRE.test(docsRepo)
-                ? docsRepo
-                : `https://github.com/${docsRepo}`;
-
-            return (
-                base.replace(endingSlashRE, "") +
-                `/edit/${docsBranch}` +
-                (docsDir ? "/" + docsDir.replace(endingSlashRE, "") : "") +
-                path
-            );
-        }
-    }
 };
 
 </script>
